+++ pydantic_resolve/__init__.py
@@ -1,3 +1,4 @@
+from .utils.collector import Collector, ICollector
 from .utils.class_util import ensure_subset
 from .utils.dataloader import build_list, build_object, copy_dataloader_kls
 from .utils.conversion import mapper
@@ -8,7 +9,7 @@ from .exceptions import (
     MissingAnnotationError,
     GlobalLoaderFieldOverlappedError)
 from .resolver import Resolver
-from .core import LoaderDepend, Collector, ICollector
+from .utils.depend import LoaderDepend
 from .utils.openapi import (
     model_config)
