+++ pydantic_resolve/utils/depend.py
@@ -1,0 +1,15 @@
+from typing import Any, Callable, Optional
+
+
+class Depends:
+    def __init__(
+        self,
+        dependency: Optional[Callable[..., Any]] = None,
+    ):
+        self.dependency = dependency
+
+
+def LoaderDepend(  # noqa: N802
+    dependency: Optional[Callable[..., Any]] = None,
+) -> Any:
+    return Depends(dependency=dependency)
