+++ tests/pydantic_v1/resolver/test_35_collector.py
@@ -1,10 +1,9 @@
 from __future__ import annotations
 from pydantic import BaseModel
 from typing import List
-from pydantic_resolve import Resolver, Collector
+from pydantic_resolve import Resolver, Collector, LoaderDepend
 import pytest
 
-from pydantic_resolve.core import LoaderDepend
 
 class SubCollector(Collector):
     def add(self, val):  # replace with your implementation
