+++ tests/pydantic_v2/core/test_field_dataclass_anno.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass, field
 from typing import Optional, List
-from pydantic_resolve.core import scan_and_store_metadata
+from pydantic_resolve.analysis import scan_and_store_metadata
 from pydantic_resolve import LoaderDepend, Collector
